Order Detail Control
- Navigation View와 Navigation Link 처리
- OrderView에 Favorite Menu View와 Menu View로 구분해서 처리
